+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Configuration.php
@@ -82,15 +82,6 @@ class Configuration implements ConfigurationInterface
                     return $v;
                 })
             ->end()
-            ->validate()
-                ->always(function ($v) {
-                    if (!isset($v['handle_all_throwables'])) {
-                        trigger_deprecation('symfony/framework-bundle', '6.4', 'Not setting the "framework.handle_all_throwables" config option is deprecated. It will default to "true" in 7.0.');
-                    }
-
-                    return $v;
-                })
-            ->end()
             ->fixXmlConfig('enabled_locale')
             ->children()
                 ->scalarNode('secret')->end()
@@ -137,7 +128,7 @@ class Configuration implements ConfigurationInterface
                 ->scalarNode('error_controller')
                     ->defaultValue('error_controller')
                 ->end()
-                ->booleanNode('handle_all_throwables')->info('HttpKernel will handle all kinds of \Throwable')->end()
+                ->booleanNode('handle_all_throwables')->info('HttpKernel will handle all kinds of \Throwable')->defaultTrue()->end()
             ->end()
         ;
 
@@ -649,38 +640,15 @@ class Configuration implements ConfigurationInterface
     private function addSessionSection(ArrayNodeDefinition $rootNode): void
     {
         $rootNode
-            ->validate()
-                ->always(function (array $v): array {
-                    if ($v['session']['enabled']) {
-                        if (!\array_key_exists('cookie_secure', $v['session'])) {
-                            trigger_deprecation('symfony/framework-bundle', '6.4', 'Not setting the "framework.session.cookie_secure" config option is deprecated. It will default to "auto" in 7.0.');
-                        }
-
-                        if (!\array_key_exists('cookie_samesite', $v['session'])) {
-                            trigger_deprecation('symfony/framework-bundle', '6.4', 'Not setting the "framework.session.cookie_samesite" config option is deprecated. It will default to "lax" in 7.0.');
-                        }
-
-                        if (!\array_key_exists('handler_id', $v['session']) && !\array_key_exists('handler_id', $v['save_path'])) {
-                            trigger_deprecation('symfony/framework-bundle', '6.4', 'Not setting either "framework.session.handler_id" or "save_path" config options is deprecated; "handler_id" will default to null in 7.0 if "save_path" is not set and to "session.handler.native_file" otherwise.');
-                        }
-                    }
-
-                    $v['session'] += [
-                        'cookie_samesite' => null,
-                        'handler_id' => 'session.handler.native_file',
-                        'save_path' => '%kernel.cache_dir%/sessions',
-                    ];
-
-                    return $v;
-                })
-            ->end()
             ->children()
                 ->arrayNode('session')
                     ->info('session configuration')
                     ->canBeEnabled()
                     ->children()
                         ->scalarNode('storage_factory_id')->defaultValue('session.storage.factory.native')->end()
-                        ->scalarNode('handler_id')->end()
+                        ->scalarNode('handler_id')
+                            ->info('Defaults to using the native session handler, or to the native *file* session handler if "save_path" is not null.')
+                        ->end()
                         ->scalarNode('name')
                             ->validate()
                                 ->ifTrue(function ($v) {
@@ -694,14 +662,16 @@ class Configuration implements ConfigurationInterface
                         ->scalarNode('cookie_lifetime')->end()
                         ->scalarNode('cookie_path')->end()
                         ->scalarNode('cookie_domain')->end()
-                        ->enumNode('cookie_secure')->values([true, false, 'auto'])->end()
+                        ->enumNode('cookie_secure')->values([true, false, 'auto'])->defaultValue('auto')->end()
                         ->booleanNode('cookie_httponly')->defaultTrue()->end()
-                        ->enumNode('cookie_samesite')->values([null, Cookie::SAMESITE_LAX, Cookie::SAMESITE_STRICT, Cookie::SAMESITE_NONE])->end()
+                        ->enumNode('cookie_samesite')->values([null, Cookie::SAMESITE_LAX, Cookie::SAMESITE_STRICT, Cookie::SAMESITE_NONE])->defaultValue('lax')->end()
                         ->booleanNode('use_cookies')->end()
                         ->scalarNode('gc_divisor')->end()
                         ->scalarNode('gc_probability')->defaultValue(1)->end()
                         ->scalarNode('gc_maxlifetime')->end()
-                        ->scalarNode('save_path')->end()
+                        ->scalarNode('save_path')
+                            ->info('Defaults to "%kernel.cache_dir%/sessions" if the "handler_id" option is not null')
+                        ->end()
                         ->integerNode('metadata_update_threshold')
                             ->defaultValue(0)
                             ->info('seconds to wait between 2 session metadata updates')
@@ -1013,25 +983,6 @@ class Configuration implements ConfigurationInterface
     private function addValidationSection(ArrayNodeDefinition $rootNode, callable $enableIfStandalone): void
     {
         $rootNode
-            ->validate()
-                ->always(function ($v) {
-                    if ($v['validation']['enabled'] && !\array_key_exists('email_validation_mode', $v['validation'])) {
-                        trigger_deprecation('symfony/framework-bundle', '6.4', 'Not setting the "framework.validation.email_validation_mode" config option is deprecated. It will default to "html5" in 7.0.');
-                    }
-
-                    if (isset($v['enable_annotations'])) {
-                        trigger_deprecation('symfony/framework-bundle', '6.4', 'Option "enable_annotations" at "framework.validation" is deprecated. Use the "enable_attributes" option instead.');
-
-                        if (!isset($v['enable_attributes'])) {
-                            $v['enable_attributes'] = $v['enable_annotations'];
-                        } else {
-                            throw new LogicException('The "enable_annotations" and "enable_attributes" options at path "framework.validation" must not be both set. Only the "enable_attributes" option must be used.');
-                        }
-                    }
-
-                    return $v;
-                })
-            ->end()
             ->children()
                 ->arrayNode('validation')
                     ->info('validation configuration')
@@ -1047,7 +998,7 @@ class Configuration implements ConfigurationInterface
                             ->validate()->castToArray()->end()
                         ->end()
                         ->scalarNode('translation_domain')->defaultValue('validators')->end()
-                        ->enumNode('email_validation_mode')->values(['html5', 'loose', 'strict'])->end()
+                        ->enumNode('email_validation_mode')->values(['html5', 'loose', 'strict'])->defaultValue('html5')->end()
                         ->arrayNode('mapping')
                             ->addDefaultsIfNotSet()
                             ->fixXmlConfig('path')
@@ -1312,17 +1263,6 @@ class Configuration implements ConfigurationInterface
     private function addPhpErrorsSection(ArrayNodeDefinition $rootNode): void
     {
         $rootNode
-            ->validate()
-                ->always(function (array $v): array {
-                    if (!\array_key_exists('log', $v['php_errors'])) {
-                        trigger_deprecation('symfony/framework-bundle', '6.4', 'Not setting the "framework.php_errors.log" config option is deprecated. It will default to "true" in 7.0.');
-
-                        $v['php_errors']['log'] = $this->debug;
-                    }
-
-                    return $v;
-                })
-            ->end()
             ->children()
                 ->arrayNode('php_errors')
                     ->info('PHP errors handling configuration')
@@ -1332,6 +1272,7 @@ class Configuration implements ConfigurationInterface
                             ->info('Use the application logger instead of the PHP logger for logging PHP errors.')
                             ->example('"true" to use the default configuration: log all errors. "false" to disable. An integer bit field of E_* constants, or an array mapping E_* constants to log levels.')
                             ->treatNullLike($this->debug)
+                            ->defaultTrue()
                             ->beforeNormalization()
                                 ->ifArray()
                                 ->then(function (array $v): array {
@@ -2344,23 +2285,6 @@ class Configuration implements ConfigurationInterface
     private function addUidSection(ArrayNodeDefinition $rootNode, callable $enableIfStandalone): void
     {
         $rootNode
-            ->validate()
-                ->always(function ($v) {
-                    if ($v['uid']['enabled']) {
-                        if (!\array_key_exists('default_uuid_version', $v['uid'])) {
-                            trigger_deprecation('symfony/framework-bundle', '6.4', 'Not setting the "framework.uid.default_uuid_version" config option is deprecated. It will default to "7" in 7.0.');
-                        }
-
-                        if (!\array_key_exists('time_based_uuid_version', $v['uid'])) {
-                            trigger_deprecation('symfony/framework-bundle', '6.4', 'Not setting the "framework.uid.time_based_uuid_version" config option is deprecated. It will default to "7" in 7.0.');
-                        }
-                    }
-
-                    $v['uid'] += ['default_uuid_version' => 6, 'time_based_uuid_version' => 6];
-
-                    return $v;
-                })
-            ->end()
             ->children()
                 ->arrayNode('uid')
                     ->info('Uid configuration')
@@ -2369,6 +2293,7 @@ class Configuration implements ConfigurationInterface
                     ->children()
                         ->enumNode('default_uuid_version')
                             ->values([7, 6, 4, 1])
+                            ->defaultValue(7)
                         ->end()
                         ->enumNode('name_based_uuid_version')
                             ->defaultValue(5)
@@ -2379,6 +2304,7 @@ class Configuration implements ConfigurationInterface
                         ->end()
                         ->enumNode('time_based_uuid_version')
                             ->values([7, 6, 1])
+                            ->defaultValue(7)
                         ->end()
                         ->scalarNode('time_based_uuid_node')
                             ->cannotBeEmpty()